Moved data processing functions into own file; removed from orion.py
